+++ src/db/mod.rs
@@ -1,0 +1,2 @@
+pub mod path;
+pub mod settings;
