+++ src/lib.rs
@@ -1,135 +1,9 @@
+pub mod db;
+pub mod types;
+
 use sysinfo::{Disks, System};
 
-pub mod database {
-    use std::env;
-    use std::path::PathBuf;
-    use rusqlite::{Connection, Result as SqliteResult};
-    use crate::{get_if_dev, Dimension};
 
-    // Determines the appropriate database file path
-    // In development: Displays relevant environment information
-    // Returns: PathBuf containing the database location
-    pub fn set_db_path() -> PathBuf {
-        let _if_dev = get_if_dev();
-
-        if _if_dev == Some(true) {
-            println!("Running in dev mode");
-            let mut _dev_path = PathBuf::new();
-            _dev_path.push(env::current_exe().unwrap());
-            _dev_path = _dev_path.parent().unwrap().join("app.db");
-            _dev_path
-        } else {
-            println!("Running in release mode");
-            // Get the path to the app bundle's Resources directory
-            let mut _release_path = PathBuf::new();
-            if let Ok(exe_path) = env::current_exe() {
-                // On macOS, the executable is typically at MyApp.app/Contents/MacOS/executable
-                // We want to store the database in MyApp.app/Contents/Resources/
-                if let Some(exe_dir) = exe_path.parent() {
-                    if let Some(macos_dir) = exe_dir.parent() {
-                        if let Some(contents_dir) = macos_dir.parent() {
-                            _release_path = contents_dir.join("Resources").join("app.db");
-                        }
-                    }
-                }
-            }
-
-            // Fallback to the home directory if we can't access the app bundle
-            if _release_path.parent().is_none() {
-                _release_path = env::home_dir()
-                    .unwrap()
-                    .join("Library")
-                    .join("Application Support")
-                    .join(env!("CARGO_PKG_NAME"))
-                    .join("app.db");
-            }
-
-            // Ensure the parent directory exists
-            if let Some(parent) = _release_path.parent() {
-                std::fs::create_dir_all(parent).unwrap_or_else(|e| {
-                    eprintln!("Failed to create database directory: {}", e);
-                });
-            }
-            _release_path
-        }
-    }
-    pub fn init_db() -> SqliteResult<Connection> {
-        let db_path = set_db_path();
-        println!("Using database at: {}", db_path.display());
-        let conn = Connection::open(db_path)?;
-
-        // Create the table only if it doesn't exist
-        conn.execute(
-            "CREATE TABLE IF NOT EXISTS UserSettings (
-            id INTEGER PRIMARY KEY,
-            item_name TEXT NOT NULL,
-            content TEXT NOT NULL,
-            created_at DATETIME DEFAULT CURRENT_TIMESTAMP
-        )",
-            [],
-        )?;
-
-        // Insert default values ONLY if they don't exist (using INSERT OR IGNORE)
-        conn.execute(
-            "INSERT OR IGNORE INTO UserSettings (id, item_name, content)
-         VALUES
-            (1, 'Default Entry', 'Initial Setting'),
-            (2, 'WindowWidth', '600'),
-            (3, 'WindowHeight', '300')",
-            [],
-        )?;
-
-        Ok(conn)
-    }
-
-    pub fn get_window_position(conn: &Connection) -> Dimension {
-        let mut my_dimension = Dimension::default();
-        my_dimension.x_position = conn
-            .query_row(
-                "SELECT content FROM UserSettings WHERE item_name = 'WindowWidth'",
-                [],
-                |row| row.get(0),
-            )
-            .expect("Read Failure");
-
-        my_dimension.y_position = conn
-            .query_row(
-                "SELECT content FROM UserSettings WHERE item_name = 'WindowHeight'",
-                [],
-                |row| row.get(0),
-            )
-            .expect("Read Failure");
-        // Return the dimensions
-        my_dimension
-    }
-    // Retrieves the currently stored entry from the database
-    // Returns: Result containing the stored content string
-    pub fn get_saved_entry(conn: &Connection) -> SqliteResult<String> {
-        conn.query_row("SELECT content FROM UserSettings WHERE id = 1", [], |row| {
-            row.get(0)
-        })
-    }
-    pub fn set_saved_entry(conn: &Connection, entry: &str) {
-        conn.execute("UPDATE UserSettings SET content = ?1 WHERE id = 1", [entry])
-            .expect("Unable to Save Entry");
-    }
-
-    pub fn set_window_position(conn: &Connection, width: i32, height: i32) {
-        conn.execute(
-            "REPLACE INTO UserSettings (id, item_name, content)
-         VALUES
-            (2, 'WindowWidth', ?1),
-            (3, 'WindowHeight', ?2)",
-            [width, height],
-        )
-            .expect("Unable to Save Window Position");
-    }
-}
-#[derive(Debug, Default)]
-pub struct Dimension {
-    pub x_position: Option<String>,
-    pub y_position: Option<String>,
-}
 pub struct Processor {
     pub name: String,
     pub vendor: String,
@@ -329,14 +203,4 @@ pub fn get_cpu_architecture() -> String {
     }
     //TODO add more architectures
     "Unknown Architecture".to_string()
-}
-
-// Determines the execution environment based on debug assertions
-// Returns Some(true) for development mode, Some(false) for release mode
-pub fn get_if_dev() -> Option<bool> {
-    if cfg!(debug_assertions) {
-        Some(true)
-    } else {
-        Some(false)
-    }
 }
