+++ src/types.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, Default)]
+pub struct Dimension {
+    pub x_position: Option<String>,
+    pub y_position: Option<String>,
+}
